Some suggested revisions
Overall fantastic explanation that walks the reader through the levels of abstraction, optimizations and implementation details.

These suggestions are mostly either typo fixes or suggestions to make the sentence easier to understand. Feel free to take or leave the at an individual level.

One or two contain a question/suggestion in parenthesis when perhaps I didn't understand if the word choice is right.

Hope this helps!